Move 'autogen' code into the method_generators module
